Combined graphql_cold_chain crate
