+++ server/graphql/Cargo.toml
@@ -16,12 +16,8 @@ graphql_core = { path = "core" }
 
 graphql_asset = { path = "asset" }
 graphql_general = { path = "general" }
-graphql_sensor = { path = "sensor" }
-graphql_temperature_breach = { path = "temperature_breach" }
-graphql_temperature_log = { path = "temperature_log" }
+graphql_cold_chain = { path = "cold_chain" }
 asset_catalogue = { path="catalogue" }
-graphql_temperature_notification = { path = "temperature_notification" }
-
 graphql_invoice = { path = "invoice" }
 graphql_invoice_line = { path = "invoice_line" }
 graphql_location = { path = "location" }
