+++ server/graphql/cold_chain/src/mutations/mod.rs
@@ -1,0 +1,4 @@
+pub mod temperature_breach;
+pub use temperature_breach::*;
+pub mod sensor;
+pub use sensor::*;
