+++ server/graphql/cold_chain/src/mutations/sensor.rs
