--- server/graphql/sensor/src/mutations/mod.rs
@@ -1,3 +1,0 @@
-mod update;
-
-pub use update::*;
