--- server/graphql/temperature_breach/src/mutations/mod.rs
@@ -1,2 +1,0 @@
-pub mod update;
-pub use update::*;
